camwaterpay+backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.36
aiosqlite==0.20.0
alembic==1.13.2
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
httpx==0.27.2
stripe==10.12.0
